func

+++ tasks/chaschin_v_sobel_operator/mpi/src/ops_mpi.cpp
@@ -45,7 +45,7 @@ std::vector<float> ChaschinVSobelOperatorMPI::PreprocessToGrayscaleWithOverlap(
   sendcounts.resize(n_procs);
   displs.resize(n_procs);
 
-  for (int i = 0; i < displs.size(); i++) {
+  for (size_t i = 0; i < displs.size(); i++) {
     displs[i] = 0;
   }
